Fix (pdfwrite) : Update CFF interpreter to handle gsubr while searching for SEAC
Bug #691680 "PDF Writer drops many accented characters"

pdfwrite makes use of a routine 'gs_type1_piece_codes' while copying fonts to determine
if a glyph is a SEAC, and therefore includes two subsidiary glyphs. This code was
updated previously (rev 10076) but on encountering a callgsubr in a CFF font, assumed
that the glyph was not a SEAC>

This patch adds support for callgsubr and continues searching glyphs which use a gsubr
to see if the gsubr is actually a SEAC type endchar.

No differences expected.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11822 a1074d23-0009-0410-80fe-cf8c14f379e6