Fix : DeviceN not permitting multiple identical inks
Bug #691740

The DeviceN checking code did not permit multiple inks to have the same name. The
comment in the code says this is not permitted, but I am unable to find any
documentation in any of the specifications in support of this, and other interpreters
(including CPSI) do not make this restriction.

Restriction removed, and also the 'continuation' procedures used for much of the colour
space parsing/validation have been given names so that future error messages use these
instead of printing the function address.

No differences expected.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11872 a1074d23-0009-0410-80fe-cf8c14f379e6