In the event that we've decided not to cache a glyph generated by FAPI, we need to make sure that zsetcachedevice2() doesn't opt to try to cache said glyph. To do that we were setting in_cachedevice to CACHE_DEVICE_NOT_CACHING before calling zchar_set_cache(), and setting it back to its previous value on return. Unfortunately, a font with a CDevProc causes the call to zsetcachedevice2() to be delayed until after execution of the CDevProc, so in_cachedevice was (probably) no longer set to CACHE_DEVICE_NOT_CACHING.
Amongst other effects, this caused some anti-aliased glyphs to be render
with incorrect, scaling as well as the potential for marking glyphs to have
empty (but valid) cached bitmaps.

Expected cluster differences: fts_23_2306.pdf at 72dpi shows some (benign)
pixel differences on the pattern filled glyphs - this is actually a behaviour
progression, as it is caused by cache being properly bypassed for an uncached
glyph, rather than the glyph being rendered uncached but the cache receiving
an emtpy glyph.

Bug 691750.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11888 a1074d23-0009-0410-80fe-cf8c14f379e6