Fix for PCL that uses the chunk allocator (which is not inherently thread safe) when NumRenderingThreads > 1. Adds 'is_thread_safe' to the gs_memory_status so that clist_setup_render_threads can wrap an unsafe allocator in a locking wrapper.
Also restore the optimizations to the chunk memory allocator to keep two lists
(one for multiple object chunks and another for single object chunks) and fixes
the chunk_free_object so that the correct list is chosen. Even if not found on
the selected list, the other list will be scanned and a diagnostic printed if
the wrong chunk list was selected (belt and suspenders). My testing with the
lj*.prn files did not show this problem.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11903 a1074d23-0009-0410-80fe-cf8c14f379e6