Revert revision 11901.
Because of the way that PCL draws bitmap fonts directly to the cache there is no
possibility of making uncached glyphs work properly. Also the code for cached glyphs is
much too forgiving and attempts to add glyphs which cannot be handled. Finally there is
no provision for type 3 fonts with non-identity matrices. Because the bitmaps in the
cache already have the scaling/rotation/shearing and clipping applied, we cannot have
a type 3 font with a non-identity matrix.

The code will be revised and recommitted.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11908 a1074d23-0009-0410-80fe-cf8c14f379e6