Fix (pdfwrite) : pointers not marked for garbage collection
Bug #691779 "SegFault with pdfwrite and more than one cid font"

pdfwrite does lazy creation of Identity ToUnicode CMaps for inclusion in output PDF
files (2 CMaps, one for horizontal and one for vertical writing). These pointers were
not marked for the garabge collector, but were stored directly in the pdf device
structure.

The CMaps were assigned to a pdfont resource type, where the pointer to the CMap *was*
marked for the garbage collector. This meant that if the pdfont resource was moved as
a result of garbage collection, the CMap could be moved as well. This left a dangling
pointer in the device structure.

If another font resource required an identity CMap then the now garbage pointer from
the device structure would be assigned. If the new font resource was moved as a result
of garbage collection, then the attempt to relocate the CMap would fail and cause a
crash.

Fixed by marking the pointers in the device structure for the garbage collector.

No differences expected.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11920 a1074d23-0009-0410-80fe-cf8c14f379e6