Due to "lazy evaluation" of softmask groups, the graphics state can change between the creation of the softmask group form, and when it is executed.
We'll now save the gstate and (most of) the ExtGState when the softmask group is
defined, and then set both before evaluting the object. A few ExtGState entries
are not easily set directly, so those are missing.

This partially fixes Bug 691157, and fixes Bugs 690351, 691348 and 690535. It
also resolves the remaining issue with 690535.

Cluster differences as follows:
Bug690535.pdf (noted above)
Bug690022.pdf, Bug690115.pdf, Bug690208.pdf, fts_26_2601.pdf and
fts_26_2603.pdf are all improved when compared to Acrobat.

A few xps->pdf conversions thru pdfwrite render differently, but they were
wrong before, and slightly differently wrong now. I believe the PDFs are
"incorrect".

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11922 a1074d23-0009-0410-80fe-cf8c14f379e6