Fix (pdfwrite) : Work around Acrobat bug with /Encoding
Bug #691809 "Regression: some PDF files produced by GhostPCL r11913 cannot be read by
Acrobat 8.2.5"

Acrobat 4 & 8 have a peculiar bug, if a type 3 font (may affect other font types, but
it seems unlikely) has a WinAnsiEncoding, and no /Differences, then text set in that
font is not displayed by these versions of Acrobat. This may be limited to glyphs
which consist of a bitmap only.

Forcing the emission of a /Differences array for each glyph used results in these
versions of Acrobat displaying correctly (even though the 'differences' are in fact the
standard Encodings).

This does seem to be quite definitely an Acrobat bug which we are working around, no
other PDF consumer seems to have a problem with these files.

Expected differences
none

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11931 a1074d23-0009-0410-80fe-cf8c14f379e6