A change to how we get the advance width of the glyph from Freetype.
Previously, we used the data in the glyph metrics structure, which
are scaled, and the undid the scaling. This is correct method for
everything except the horizontal and vertical advance distances.

We now get the two advance distances from the glyph structure,
and tell Freetype not to scale them, this avoids the rounding
errors that occur due to Freetype's fixed point number
format. It also saves the (less significant) error when we
previously unscaled the value, and saves a couple of
fairly large double precision calculations.

This does cause a number of changes in the regression suite,
the vast majority are pixel or so differences in glyph
positions. A very few are improvements.

Bugs 691778 and 691777

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11943 a1074d23-0009-0410-80fe-cf8c14f379e6