pdfwrite enhancement : DSC output
Add some more comments, and re-arrange a few to make better sense. Initial tests
indicate the output of the code with ProduceDSC does not introduce any new errors.

Still to do: See if we can re-order the output so that we can write the %%Pages: comment
in the header with the correct number of pages. This requires us to defer writing the
header and ProceSet until the end of the job.

No differences expected.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11950 a1074d23-0009-0410-80fe-cf8c14f379e6