Bug #691835
Temporary work-around. It is possible to get into gsicc_init_device_profile with the
graphics state pointer (pgs) being NULL. This happens if ps2write or pdfwrite convert
a masked image into a regular image.

Because pgs is dereferenced without checking for NULL, this causes a crash.

Its not immediately clear to me if I need to modify ps2write/pdfwrite so that this
condition doesn't occur, but for now, if pgs is NULL we return from the routine in order
to prevent a crash.

No differences expected

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11961 a1074d23-0009-0410-80fe-cf8c14f379e6