Fix ps2write/pdfwrite : Up date the show enumerator with the cached character width
When capturing a character bitmap on the first time it is rendered, pdfwrite and
ps2write get the width of the glyph in the cache device. This is then used to update
values in the show enumerator, and these values are used as the Widths entry for the
glyph.

When imaging an already cached character, however, the enumerator values are not updated
which can lead to garbage (uninitialised data) being written as values in the /Widths
array. This patch takes the value from the cached character 'wxy' and updates the show
enumerator 'wxy' with it. This is then used as the value for the /Widths array.

Bug #691836, zweiseit01d4-1.pdf, limitcheck --nostringval--

No differences expected.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11969 a1074d23-0009-0410-80fe-cf8c14f379e6