Add new top level GhostPDL makefile targets, pcl-lib, ls-lib, svg-lib, xps-lib and all-lib.
These build the same as pcl, ls-product, svg, xps etc, except the final
product is a library that offers pl_main() as its top level entrypoint.
The sole piece of code that is NOT included in the libs that is included
in the exes is a 1 line 'main()' veneer that just calls pl_main.

No expected cluster differences.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11973 a1074d23-0009-0410-80fe-cf8c14f379e6