More DSC-compliant ps2write work.
When emitting strings for ps2write, write them as hex strings instead of regular
strings, if the ps2write output is to be DSC-compliant, break the hex strings every
255 bytes or less. Prepend and append newlines to the hex string when emitting
DSC-compliant PostScript to ensure that preceding or following data doesn't cause the
line length to exceed 255 bytes.

No differences expected as ps2write is not currently regression tested.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@11994 a1074d23-0009-0410-80fe-cf8c14f379e6