Added the source code of the CUPS libraries to the Ghostscript source package
Like with zlib, libpng, libjpeg, ... it is now possible to
build the complete Ghostscript (including the CUPS Raster output
device) on systems where there is no CUPS library available.

Note that currently only the CUPS library source code is added, the
results of the ./configure run in the Makedefs and config.h files, and
the Makefiles are stripped to only build the libraries and nothing
else. The CUPS ./configure command line which was used to obtain
makedefs and config.h was

./configure --enable-static --without-shared --disable-dbus --without-php \
--disable-slp --disable-gssapi --disable-ldap --disable-ssl \
--disable-cdsassl --disable-gnutls --disable-openssl --disable-pam \
--disable-dnssd --disable-launchd

This way the built CUPS libraries do not need any system services or
hardware interfaces of Linux. In addition to standard libraries they
use only libz, libpng, libjpeg, and libtiff. These libraries
Ghostscript needs anyway and copies of their source code also ship
with Ghostscript.

What is still missing is the integration into the build system of
Ghostscript and also the operating-system-independent base/cups.mak
file (like base/png.mak etc.).

No changes expected in the built Ghostscript as the code is not yet
integrated in the build system.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@12017 a1074d23-0009-0410-80fe-cf8c14f379e6