Make sure the X11 drivers get removed from the DEVICE_DEVS declaration if the --enable-dynamic option is give to configure. Also, fix up the the non shared library builds, so they will still work with --enable-dynamic  (only really useful for debugging).
Bug 691905.

Reorganise and correct the names of the devices specified in
configure.ac so that the options --with-drivers=[FILES,PRINTERS,X11]
actually work.

Finally, make the "make debug" compiler options originate from
configure, instead of being hard coded into the Makefile. A first
step in making some of useful debugging features of modern compilers
available when we can.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@12049 a1074d23-0009-0410-80fe-cf8c14f379e6