Add code to define cbrt, strcasecmp and stncasecmp when compiling for Windows, as these are not present in the MSVC run-time.
Also alter the initialisation of the gs_cups_device structure. Previously this set three
matrices by including empty braces '{}', but MSVC throws a compilation error on this.
Altered to {0x00} which is not quite the same, but it compiles (with a warning from gcc)
and is no worse than leaving the members uninitialised.

No differences expected.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@12060 a1074d23-0009-0410-80fe-cf8c14f379e6