While searching for performance on an ARM based dev board, I developed 2 new versions of the skip_white_pixels macro. The first uses native int sized checking, and should be faster on long run matching. I have as yet been unable to find any files that show this expected speedup.
The second should be faster on architectures (such as ARM) where the tradeoff
in the test is different (and where pointer arithmetic is cheaper than array
indexing).

Both versions are disabled by default until the case for their existence has
been proven, but are committed here to avoid bitrot.

No cluster differences expected, as the default code does not change.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@12085 a1074d23-0009-0410-80fe-cf8c14f379e6