Fix (pdfwrite) : permit symbolic fonts in PDF/A output
Bug #691877 "Invalid text when converting to PDF/A"

TrueType fonts were being emitted as non-symbolic when PDF/A output was selected. This
was causing the text to be incorrect, because the fonts actually are symbolic and
Acrobat was unable to apply the correct decoding.

This will result in more text being unsearchable/copyable but the rendering result is
correct, which is more important.

No differences expected, PDF/A output not tested on the cluster.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@12086 a1074d23-0009-0410-80fe-cf8c14f379e6