Add new 'Profile' configuration to Visual Studio targets. This builds objects/binaries with debugging symbols enabled, but otherwise full optimisation.
The changes are entirely within the VS files, except for a few small tweaks
to the MSVC makefile. These changes should mean that it is possible to use
the makefiles in a way that doesn't rely on recursive make calls (even though
we leave the 'debug' targets in for now that do use that).

No cluster differences expected or shown in testing.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@12087 a1074d23-0009-0410-80fe-cf8c14f379e6