Fix (pdfwrite) : Always write a ToUnicode CMap for TrueType fonts
Bug #691907 "PDFs with TrueType fonts from Windows PostScript files not searchable"

Patch from SaGS.

When choosing whether to embed a ToUnicode CMap, always embed if the font type is
TrueType. This is because TT fonts are always embedded as symbolic, and we now no longer
add Encodings to Symbolic TT fonts (violates spec), which results in PDF files where the
text is unsearchable and cannot be successfully copy/pasted in the absence of a
ToUnicode CMap.

It is possible that the ToUnicode CMap is incorrectly generated, but this is no worse
than having no CMap at all.

No expected differences since ToUnicode CMaps are not used in rendering.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@12088 a1074d23-0009-0410-80fe-cf8c14f379e6