Fix ps2write : Don't depend on the PS file system for resources.
ps2write needs up to 4 resource files which are normally stored in the PostScript ROM
file system. However on PCL and other languages these resources are (obviously) not
present, so the output is incomplete.

Move the relevant files from PostScript into 'C' as header files. Remove the files from
the PostScript file system to save ROM file system space. Put the removed files into
the gs/lib folder for safe-keeping. Add the .h files as dependencies for gdevpdfu.c in
the build system so that changes to these files cause a rebuild.

No differences expected.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@12093 a1074d23-0009-0410-80fe-cf8c14f379e6