A set of changes to make it easier to use the UFST with Ghostscript and to get us closer to having the option of the UFST handling font types other than the Microtype fonts.
The UFST build, using the default COMPILE_INITS=1  will now include
the Microtype FCO files in the rom file system along with the other
initialisation files (it does not currently strip out the normal
Postscript fonts in Resource/Fonts), and setup default values for
the various paramters required to have Ghostscript use the
Microtype fonts (these can still be overridden on the command
line). For the build with COMPILE_INITS=0, the default paths
will use the path to the UFST you specify for configure, or
nmake on Windows.

Included in this revision are changes to prevent the UFST version
5.x (or less) from trying to handle non-Microtype fonts, and allow
it to safely fall back to the Freetype (or a.n.other FAPI plugin).
UFST 6.2 or better will try to handle all fonts types, but this is
not yet reliable.

Currently, this code only replaces the UFST memory management on
UFST 6.2 and above (this is purely because I haven't time to test
this capability on UFST 5 yet).

A further change will be committed shortly to the UFST 5 sources
to force the UFST to use the Ghostscript stream IO instead of the
standard lib file IO on the Windows builds (this has been in place
for the Unix/Linux UFST builds for some time).

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@12099 a1074d23-0009-0410-80fe-cf8c14f379e6