Work on the pattern transparency clist code.
A squashed and rebased patch based on the following svn commits.

Author: robin <robin@a1074d23-0009-0410-80fe-cf8c14f379e6>
Date:   Tue Apr 19 12:40:00 2011 +0000

Various fixes to the transparency clist branch to correctly use
just the subrectangle of transparency tiles that is actually present.

Getting this in before the Great Git Change this afternoon.

git-svn-id: http://svn.ghostscript.com/ghostscript/branches/patt_trans_clist_gs@12407 a1074d23-0009-0410-80fe-cf8c14f379e6

Author: ray <ray@a1074d23-0009-0410-80fe-cf8c14f379e6>
Date:   Fri Apr 8 16:56:48 2011 +0000

A couple of minor fixes that help things along, get past compile errors.

git-svn-id: http://svn.ghostscript.com/ghostscript/branches/patt_trans_clist_gs@12379 a1074d23-0009-0410-80fe-cf8c14f379e6

Author: mvrhel <mvrhel@a1074d23-0009-0410-80fe-cf8c14f379e6>
Date:   Wed Feb 9 21:38:24 2011 +0000

Initial commit of work so far on getting the pattern transparency
clist code working. Much of this comes from work that Ray and I did
for customer 532. The remaining portion of the work is primarily
in pdf14_tile_pattern_fill in gdevp14.c, where we need to make
sure that the group that we are going to push is the intersection
with the rect from trans pattern since we only use the part that we
drew into in the pattern creation and not the whole group buffer. I
had not worked on this in a few weeks so I don't have more details
right now. I will spend a day on it this week to see exactly what
remains to be done.

git-svn-id: http://svn.ghostscript.com/ghostscript/branches/patt_trans_clist_gs@12137 a1074d23-0009-0410-80fe-cf8c14f379e6