Fix Bug #691918
Update the Unicode decodings applied to TrueType fonts to match the latest Adobe glyph
list. Fixes some problems with incorrect mappings and adds numerous new mappings. A
similar but less extensive change is made to the FCO_Unicode mappings as well.

Thanks to SaGS for the work on this problem.

No differences expected as these are only used for ToUnicode CMaps.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@12139 a1074d23-0009-0410-80fe-cf8c14f379e6