Fix : colour space handling bug and improved handling of broken ICC space
Bug #691941 "Interpretation of PDF aborts with /typecheck"

The PDF file in the specimen contains an invalid colour space of the form:

/Indexed [/ICCBased <</N 1 /Alternate [/Indexed /DeviceRGB 255 7 0 R]>>] 255 7 0 R]

The number of components in the ICCBased specification is incorrect, as the profile has
3 channels. This was not detected previously. Falling back to the /Alternate we see
that we have a /Indexed space depending on a /Indexed space which is also invalid, but
we choose to ignore this.

There was also a bug in the colour space handling for ICCbased spaces which caused a
typecheck if the alternate space was a /Indexed space.

Finally, the PDF interpreter is updated so that when given a stream as the data source
for a /Indexed space it reads and returns a string which is the greater of the declared
size of the stream, or the calculated size required, given the number of components.
Previously we always returned the calculated size, which was too little in this case
as the number of components in the ICCBased space is incorrect.

With these changes the (invalid) specimen file runs to completion.

No differences expected.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@12140 a1074d23-0009-0410-80fe-cf8c14f379e6