Tiny tweak to the in-development halftone code; currently only code for handling monochrome halftoning has been implemented, but the code does not recognise this and enables itself for all output depths. This leads to unexpected SEGVs when run with non-monochrome outputs (such as for example pgmraw rather than pbmraw).
This fix simply changes the code to only be used in the case of monochrome
output. This vastly reduces the number of SEGVs given by the new code, and
makes cluster testing of other tweaks much easier.

This is only a temporary measure as the halftone code is intended to cope
with non-monochrome cases too.

No cluster differences expected as this is on an untested branch.

git-svn-id: http://svn.ghostscript.com/ghostscript/branches/halftone@12155 a1074d23-0009-0410-80fe-cf8c14f379e6