Fix 2 causes of SEGVs in gximono.c, both in the landscape code.
Firstly, we were being entered with a count value larger than sixteen.
For instance if we had width values of 4 4 4 3 4 then we'd get a count of
19, and overflow buffers. We fix this by copying the widths to a local array
and truncating the last (or first depending on the direction) element to fit.
This enables us to simplify the inner loop.

Secondly, in the case of extreme upscales (more than a factor of 16) we could
complete one flush operation, and still have enough in the buffer to flush
again. If we forget to flush the buffer can overflow the next time around.
The fix is simply to repeatedly flush until we have less than a buffer full
left.

No cluster differences expected as on an untested branch.

git-svn-id: http://svn.ghostscript.com/ghostscript/branches/halftone@12159 a1074d23-0009-0410-80fe-cf8c14f379e6