Fix DO_FILL_RECT_BY_COPY_ROP code, and enable it by default.
The only thing wrong with the code before is the case when strip_tile_rectangle
is called with both color0 and color1 being gx_no_color_index. That translates
to rop=0xAA (i.e. D - no change). This is actually a special case that means
it's really doing a copy_color operation. We handle this by punting in the
same way as the old code used to.

No cluster differences expected.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@12182 a1074d23-0009-0410-80fe-cf8c14f379e6