Fix bug reported by "new customer feb 2011", whereby gs 8.71 on an embedded ppc platform is getting the colours on an image in a pdf wrong.
Debugging shows that in gs_indexed_limit_and_lookup we take a floating point
value, clip it, convert it to an int and use it to lookup which colour to use.
On the reference x86 run we have a value of 1 (0x3f800000 as an IEEE float).
On the ppc we have 0.999999 (0x3f7fffff as an IEEE float). This tiny difference
results in values of 1 and 0 respectively when converted to the int, giving
the wrong colour.

The fix here is to add a small epsilon before conversion.

A quick experiment in adding 0.5 rather than epsilon shows worse results.

15 cluster differences in testing, none that actually survived a bmpcmp.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@12205 a1074d23-0009-0410-80fe-cf8c14f379e6