Fix rop operation on plib device. Previously, I'd disabled get_bits_rectangle on the buffer device as the data is normally in the format we need it in anyway, so it's a NOP. Unfortunately it's needed for rop operation, so reintroduce it.
To avoid infinite loops, we have to cope with GB_RETURN_POINTER. This is
easy to add to the gdevmpla.c device, but it's less clear that adding it
into the mem device is the right thing to do. We therefore introduce a
shim function to cope with GB_RETURN_POINTER with the mem device.

No cluster differences expected as this is disabled by default.

Testing shows that the planar device is now very close to the non planar
equivalent.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@12229 a1074d23-0009-0410-80fe-cf8c14f379e6