Added the ability to specify bmpcmp in addition to a normal clusterpush.pl operation.  Both commands will be queued in the correct order.
Examples:

  ./clusterpush.pl gs bmpcmp
  ./clusterpush.pl bmpcmp gs pcl xps ls

Note that the order of the options is not signficant.

The command line parser for clusterpush.pl changed signficantly with this
revision.  It should be backwards compatible with the previous version
but it's possible that subtle differences exist.  If a clusterpush.pl
command line behaves differently than you expect please open a bug.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@12230 a1074d23-0009-0410-80fe-cf8c14f379e6