Add new debugging define to gximono.c, PACIFY_VALGRIND.
This enables various small tweaks in the code that stop valgrind throwing
errors. We believe that all the errors thrown are false positives, but
we enable this define anyway until we've sorted the current indeterminisms.
We'll disable it in a few days when we have solved the problems and check that
it really doesn't cause any more.

Cluster results unknown; probably no change. If this solves indetermisms
then we'll need to understand why.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@12255 a1074d23-0009-0410-80fe-cf8c14f379e6