Change to gsroprun1.h to avoid over/underreading the source/texture buffers.
Given a valid byte range we expand that to the smallest enclosing CHUNK range
and guarantee never to access out of that range. Previously we could read
one CHUNK before/after it.

If this is a problem, simply ensure that CHUNK is byte rather than int on
your platform. This now behaves better than the original code which would
access one byte before/after the defined range.

No cluster differences seen in testing.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@12262 a1074d23-0009-0410-80fe-cf8c14f379e6