Disable PACIFY_VALGRIND in gximono.c. This define is intended to enable extra code that can be performed so as to ensure that valgrind doesn't report false positives. As such, disabling it should have no adverse effects.
Unfortunately, it seems that in the portrait case, if we don't blank the
threshold array before we run, we get diffs. I have therefore taken this
memset out of the PACIFY_VALGRIND case and forced it to always happen.
This probably points to a bug and should be investigated properly.

No cluster differences expected.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@12264 a1074d23-0009-0410-80fe-cf8c14f379e6