Refactor the Windows display device code to remove the duplication of elements in pl_main.c.
This introduces a new API call pl_main_aux which accepts an additional parameter
containing a pointer to a display_callback structure if the display device is to
be used. This  should be NULL if the display device is not used. If the display
device is selected and the callback structure pointer is NULL then an error is raised.

No differences expected.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@12293 a1074d23-0009-0410-80fe-cf8c14f379e6