Add new copy_plane device procedure; arguments as for copy_color with an extra int to identify which plane we want to work on.
The call effectively says "Copy this pixmap into the specified plane of this
device". It is envisaged that it will only be defined for planar devices.

For now clients can detect whether or not to use this based on whether the
function pointer is non NULL. If need be (to cope with forwarding devices)
we can add a dev_spec_op.

No cluster differences expected.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@12329 a1074d23-0009-0410-80fe-cf8c14f379e6