Initial import of LittleCMS v2.1 into the Ghostscript/GhostPDL build, with simple integration. LittleCMS version 1 is still used by default, so no cluster differences expected.
The version of little cms imported was taken from the lcms git repository,
commit 314bc7201d340ab303b36a0ade2c37cf40b83a3b, and then a couple of
small tweaks were applied:

 * lcms2 currently offers no way to get the input/output format for a
    transform, so I add 2 functions (cmsGetTransformInputFormat and
    cmsGetTransformOutputFormat) to do this.

 * lcms2 currently offers no easy way to redirect malloc/free operations
   away from it's internal implementation, so we add a define (LCMS_USER_ALLOC)
   to allow us to be able to define our own functions. This mirrors the
   work done in lcms1 by Michael, I believe.

The choice of which version of lcms to build with is made by the WHICH_CMS
define option in the makefile. Currently this is set to lcms by default,
but can be changed to lcms2 to allow lcms2 to be used instead.

To make this work, various small tweaks have been required. Firstly, we
rename gsicc_littlecms.h to gsicc_cms.h, reflecting the fact that this is
not littlecms specific - rather, it defines the interface that any cms
would have to provide in order to work with Ghostscript/GhostPDL.

Secondly, we rename gsicc_littlecms.c to gsicc_lcms.c to make the makefile
macros easier.

We add a gsicc_lcms2.c file (derived from copying and editing gsicc_lcms.c)
that interfaces with lcms2.

Currently lcms2 builds and appears to run without crashing, but gives
differing results in some cases. This will need to be solved before we can
swap over to using lcms2 by default.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@12340 a1074d23-0009-0410-80fe-cf8c14f379e6