Bug 689093 #17/ SVN damage: Protect text-"$Id$" from expansion, part #1/2: script generating HTML
If a commit log contains the text "$Id$", or some other text that
looks like a SVN keyword, toolbin\split_changelog.py outputs it as-is
to HTML and later SVN expands it as a keyword. The result is the text
displayed differs. This patch extends the substitutions done for
character entities to include "$" => "&#36;"; there won't be any
"$"-as-text, and thus no "$Keyword..$"-as-text, anymore.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@12360 a1074d23-0009-0410-80fe-cf8c14f379e6