Bug 689093 #18bis/ SVN damage: Replace ijs\ijs_spec.pdf, with a good copy.
This file is currently damaged in the SVN Repository. The damage seem to have
happened during the conversion of the Repository from CVS to Subversion. This copy
of the file comes from the last CVS version available; the copy in the 1st SVN
checkout I have is already damaged. Note the "/CreationDate" inside the pdf is the
same as in the current (damaged) version and the last change to this file is very
old, so this file's substantive content did not change after the migration to SVN.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@12362 a1074d23-0009-0410-80fe-cf8c14f379e6