Resolve a crash to due an unexpected CMap object type.
The code was not checking that the CMap object was a string or an array (of
strings) before trying read bytes from it. If the CMap actually turned out
to be an unexpected object, such as an integer, it would cause a crash.

Credit to Ken Sharp for the patch.

Bug 692124

No cluster differences expected.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@12371 a1074d23-0009-0410-80fe-cf8c14f379e6