Change lcms2 memory handling and update lib.
As of this commit, we can (and do) run a completely unchanged library.

This includes our tiny tweak to be able to get the input/output formats for
a given transform. Also update our interface code so that we use the plugin
mechanism to redirect malloc/free etc through us rather than hacking the
code direct. Thanks to Marti Maria for both of these.

No cluster changes as this code is not tested. CLUSTER_UNTESTED.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@12372 a1074d23-0009-0410-80fe-cf8c14f379e6