Fix (pdfwrite) : Not using ToUnicode CMaps under some conditions
Bug #692119 "Cannot copy text from Ghostscript generated PDF/A document"

The PDF interpreter converts ToUnicode CMaps into GlyphNames2Unicode dictionaries, to
do this it uses the Encoding to convert the character codes into glyph names. Of course
for CIDFonts we don't want to do this, and so when there is no Encoding we instead
put the Unicode value directly into the dictionary using the CID as an index.

If the font has no Encoding (which is optional in PDF) then we do the same for a
regular font. However, by the time pdfwrite sees the font we have had to add an Encoding
as its not valid for us to define a regular font with no Encoding.

So pdfwrite converts the character code into a (bogus) glyph name, and then tries to
look up that glyph name in the GlyphNames2Unicode dictionary. Obviously this fails.

I've extended the 'decode_glyph' call so that it takes both the glyph name and the
character code, for CID fonts the character code is always -1. If the glyph name can't
be found in the GlyphNames2Unicode dictionary then we try to use the character code
instead.

This works for this test case, and I don't think its likely to produce worse results
than the old code. Regression tests show no differences, but since they don't test
ToUnicode CMaps they weren't expected to. My own tests seem to be OK but I can't say
I'm terribly sure about this one.

Expected Differences
None

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@12374 a1074d23-0009-0410-80fe-cf8c14f379e6