Fix (colour handling) : crash when executing 'stop'
This is part of the fix for bug #692116. The crash in that report is caused because the
continuation procedures for the colour handling are sometimes pushed with a
push_mark_estack where the opproc has a value of 0. This is not valid and I always
intended to fix it, but forgot.

This changes all uses of push_mark_estack so that the opproc is set to a valid routine
which does nothing, thus solving the crash and exposing the real problem. The only
reason we get a crash is because we are executing stop as part of the error handler,
because currentgray returned an error.

No differences expected.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@12382 a1074d23-0009-0410-80fe-cf8c14f379e6