Add a fallback check for libpaper support.
Older versions of autogen appear to have a bug which causes the AC_CHECK_LIB
macro to mis-identify the libpaper development library as being present when
it is not.

So add a second check check based on the presence of paper.h using
AC_CHECK_HEADER().

No cluster differences expected.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@12401 a1074d23-0009-0410-80fe-cf8c14f379e6