New tiffscaled8 and tiffscaled24 devices.
Add new tiffscaled8 and tiffscaled24 devices, copied and modified from
tiffscaled. These output greyscale and 24bit rgb instead of tiffscaleds
mono output. MinFeatureSize is ignored for these devices as it's meaningless
for contone output.

Error Diffusion is still done, but is almost certainly a waste of time - it
was just simpler to make the code work this way. If performance is an issue,
we can remove that later.