Fix memory overwrite in png devices.
In my fiddling with the png devices the other day to add downscaling, I
omitted to extend the size of two of them (png16m and png48). This meant
that png16m was overwriting 8 bytes of memory after its device structure.

Fix this to solve a SEGV when DownScaleFactor is used.