Fix for bug 692081. Antialiasing patterns results in white holes.
gx_default_copy_alpha simulates alpha by weighted average of RGB values.
When the alpha value is zero, the current implementation reads the current
background value, and just writes it back. This has the effect of making
any device that watches what devices get written to think that that pixel
is solid. (One such device is the pattern tile device where it collects what
pixels have been marked, and what haven't).

The CORRECT way to do this is to stop the line accumulation at that point,
flush everything to there, then start again a pixel later.

Doing this solves the bug, and causes no cluster changes.