Revert SVN commit 12934, ref count checking before free in gxclrast.c
A couple of weeks ago, I applied a wild stab in the dark to solve a problem
that I cannot now locate. Having discussed it with Michael it seems that this
was (at best) just masking another problem (the incorrect ref count
initialisation of compositors). This has been fixed, so my faulty attempt
can be backed out.

We leave a debug only check in there to catch this situation. This should be
replaced by an assert when we have such things.