Ensure optimization flags get propogated for lcms
An inconcistency between definitions in the Unix and Windows builds meant
that on Windows several compiler flags (including that for optimization)
was not being used when buliding the lcms source files. The confusion
stems from Ghostscript requiring /Za to compile, and lcms failing to
compile with /Za.

Bug 692173.

No cluster differences expected.