Stop compiler turning for loop into memset in halftoning code.
The msvc compiler cleverly spots that a for loop can be turned
into a memset. Unfortunately it can't know that the values for
which the loop is called are normally so small that the calling
of the function costs more than simply doing the stores.

The fix is to cast the pointer to which we are storing to be
volatile. This saves a significant chunk of runtime for:

 pcl6.exe -sDEVICE=bit -r600 -o null: -dLeadingEdge=3
             cicero_call_CRF03-all-in_adobe-8_358p_xNuv288.xl

No cluster differences expected.