Add new .strokeafterfill ps operator/graphics state bit.
We add a new .strokeafterfill ps operator that takes a bool, and sets it in
the graphics state. The intention is that this will allow us to spot the
case where we are executing the stroke part of a combined 'fill and stroke'
operation (as required by pdf). This should allow us to solve the
remaining problems with the pdf14/knockout group.

No cluster differences expected.