Change error return to remove confusion over meaning.
When we encounter a font whose charstring has been replaced with a PS
procedure, we were returning FT_Err_Invalid_File_Format, and using that
as a hint we should try to execute it as a PS procedure. This
turns out to be problematic as it is also the error return generated by
Freetype for an invalid charstring. In the case of an invalid charstring
we should fall back to the notdef, not attempt to execute the object as
Postscript.

So, change the error return value that *we* generate, and subsequently
check for, to FT_Err_Unknown_File_Format, which FT should never,
normally, generate during glyph interpretation. This means, in this
case, we correctly identify a failed attempt to interpret a charstring,
and use the notdef fallback - instead of, eventually, throwing an
error.

Bug 692176.

No cluster differences expected.