Quiet valgrind warning, and remove misleading init.
When run with valgrind on:

   gs -sDEVICE=pbmraw -r300 -dMaxBitmap=100000000 -o out.pbm Bug690189c.pdf

Valgrind complains that params.GrayBackground is used without being
initialised. We fix this here by amending gs_trans_mask_params_init.

In so doing, we note that in zbegintransparencymaskgroup we were setting
params.ColorSpace, only to call gs_trans_mask_params_init and have it
overwrite it. To avoid confusion we therefore remove this needless init.

No cluster differences seen in testing.