Fix clipping problems with composited pattern clists.
When using a clist to fill a clipped area with a pattern, the current
code can drop the clipping device out of the chain. This change solves that
problem.

When we create a new compositor, we generally want that compositor to become
the top device in the chain, so all new drawing commands go to it. In the
pattern clist case however, we don't create a new compositor - we reuse an
existing one. It's important in this case *not* to take this as our new
target, otherwise the higher devices in the chain (such as the clipper
device) are lost.

This was shown when rendering pattrans_big.pdf at 300dpi banded to the pkm
device.