Put back default map_rgb_color forwarding.
At some point around the creation of this branch (which corresponded to our
move from SVN to git) we lost a line in gdevnfwd.c that sets up a default
forwarding for map_rgb_color. We don't know why it went, so we're putting it
back in. This probably hasn't caused any problems as it's a deprecated
function now.