Solve windows debug build crash on exit with cups device.
It seems that the cups device takes it upon itself to close the output file
rather than leaving it for the ghostscript core to close as every other
device does. This upsets the windows DLL file handling.

The solution is simply to remove the lines that close the file. I've
checked both with Till and Ken, and they concur. Thanks.

No cluster differences expected.