Retain or bias the reference count of the rop texture device.
The image code was using the rop texture device without any reference
counting, explicitly freeing the device when it was done.  That was
fine if the device was the only device in the chain, however, linking
in other devices could lead indirectly to unintended freeing of the
rop device.  A crash was seen in the the XL test file C705.BIN but
only with halftoning at particular resolutions like 72 dpi.  We did
not study why the crash was limited to particular resolutions and
devices.