Use defined default page size for nullpage device.
The nullpage device previously used a 1"x1" default page size, this changes it
to use the same default setting as other page devices (this can be Letter or
A4, set in a makefile).

Also, add some lower bounds checking to gslp.ps so that it will error out if
the page size is too small to hold at least one line of text.

No cluster difference expected.