More PCL type 3 improvements
Add code to initialise the returned character code, so that the charproc
is assigned the correct character code. Further test for PCL fonts in
pdf_text_set_cache and don't try to use a show enumerator to get the
'output_char_code' when we are doing PCL, instead use the one from the
text enumerator 'returned' structure. Normally this would not be valid
yet as this is filled in after we've completed the 'show', but we
set this before startig the show, from the text string, so that it will be
valid.