More type 3 changes
Don't scale the CTM by 100 (done for FreeType) when handling PCL fonts
in install_charproc_accum, set the boolean to complete_charproc_accum so
that we don't 'undo' the factor of 100 scaling when the font is PCL.

Add code to set_charproc_attrs to determine whether this is a 'scale 100'
(ie PostScript) type 3 font or not, if its not then don't undo the scaling
by 100 of the CTM.

When accumulating a chraproc, before setting the CTM to identity matrix
also set the current point to 0,0, which ensures that that the current point
doesn't get baked into the character description. Also invalidate the
'char_tm' txy_fixed_valid member of the graphics state, this will force
a recalculation of char_tm using the new identity matrix.