Type  PCL fonts again.
Put back the matrix scaling in pdf_text_set_cache, even though the matrix
shoudl always be the identity here when running PCL. Best to be safe.

set_charproc_attrs emitted a 'd1' setcachedevice, but didn't check if
the glyph was flipped. For PCL this led to ury being less than lly, and
so the glyph was elided. Added check to make sure these are correct. This
required removal of 'const' from an arry as well.