Add LAND_BITS define to landscape mono halftoning code.
The monochrome halftone thresholding code relies on collating incoming
scanline data into 'scancolumn' buffers. When we have 16 bytes in these
buffers, they halftoned and flushed through copy_mono.

copy_mono would prefer to work with longer than 16 byte runs, so we
remove the hardcoded 16s from throughout the code and replace it with
a #defined value. A side effect of this is that the code becomes
slightly clearer.

Sadly, it seems this doesn't give the expected speedups; testing here
indicates that both 32 and 128 run slower, possibly due to cache effects.

Nonetheless we commit the code in case inspiration strikes us later on.

No cluster differences expected.