More type 3 font madness
Use the device level 'PS_accumulator' flag in various places instead of
the more kludgy test against penuym->pte_default being NULL.

If we don't get a glyph name back from the interpreter (PCL) then invent
one instead of giving up with an error.

If we are not a type 3 accumulator, then don't undo the factor of 100
scaling applied to the device width and height, we only do that for PS.

Add a routine to return a special 'initial' matrix during the course of
type 3 accumulation. The PCL stick font uses this to set the line width
and we need to account for various PS/PDF scaling which will otherwise
be ignored.

Make sure we don't try and accumulate a charproc when its being run for a
charpath operation.