LCMS (v1) optimisations for CachedXFORM etc.
Optimise CachedXFORM (and related functions) by using a 'template' header
file and repeatedly including it with different options to generate
different specific versions of the code.

This gives a 10% improvement on 568.40345_VO_nr_3_vpeSR18.xl (a file where
most of the time is spent in clipping).

No expected cluster differences.