Change PSI so -Z! works by default in debug builds.
Previously, the -Z! option (to show the name of each PS operator as it is
executed) has only been available in DEBUG builds that specifically use
the DEBUG_TRACE_PS_OPERATORS option. This change ensures that it is available
in ALL debug builds, and in any release build where the
DEBUG_TRACE_PS_OPERATORS symbol is predefined.

The reason for historically having it as a separate option was because
checking for whether the ! debug flag was set on every operator was felt
to be too large an overhead. I have arranged the code so that this overhead
is minimised now - in debug builds we only get the extra overhead when
the -Z! is specified on the command line.

In normal release builds there is no change to the overhead. In release
builds with DEBUG_TRACE_PS_OPERATORS, the same overheads apply as in
DEBUG builds (that is a small additional overhead when -Z! is not used
and a larger one when it is).

No cluster differences expected.