LCMS (v1) Optimisations for CachedXFORM etc.
Optimise CachedXFORM (and related functions) by using a 'template' header
file and repeatedly including it with different options to generate
different specific versions of the code.

This gives a 10% improvement on 568.40345_VO_nr_3_vpeSR18.xl (a file where
most of the time is spent in clipping).

No expected cluster differences. I've just done a full test with a bmpcmp
and it showed no differences, but in light of what happened last time I
committed this, I am not going to use the magic rune to stop the cluster
rechecking it on commit.