Add a new font type 'ft_GL2_stick_user_defined' and use it as the FontType of the HP/GL2 stick font. Modify code throughout to treat it the same as 'ft_user_defined', ie a type 3 font.
Allow pdfwrite to attach a type 3 'PDF font' to a stick font (normally the
types are required to be the same).

Alter the way we calculate the 'default' device matrix during stick font
accumulation so that we get better stroke widhts.

Now that we cna identify a stick font, check for anamorphic scaling before
accumulating the glyphs. If anamorphic scaling, then *don't* treat this as
a font. THis is because the stroke width will be inconsistent.