Change in device ICC profile handling
This is the major portion of the code needed to achieve object dependent
color management.  This fixes the problems that existed in the
previous code with the device parameters and introduces an array of
ICC profiles in the device structure.  The code was cluster pushed and
showed some very minor differences in a couple files but they appear to be
OK with bmpcmp.  I still need to do further testing to verify that all the
functionality is correct (e.g. make sure setting the text profile properly
affects the text only).  In addition, the rendering intent options need to be
implemented.

I also need to check that nothing was broken with respect to MT
rendering and some of the devices that are not tested with cluster
pushing (e.g. the display device and the x11alpha device).