Fix for MT crashes in lcms
It turns out that lcms is can not safely share
profiles between threads when creating links.  This
was fixed by placing a mutex lock around the profile
creation.  It also turns out that lcms can not
use the same link to do transformations between threads.
This was solved by disabling the shared cache between
the threads.  Finally there was a race condition in
reference counting the device ICC profile in the pdf14
device.  Also, this commit includes a fix for a memory
leak in the pdf_mask object in gdevp14.