Fix Bug 692216. Remove zero height traps from shadings.
The simplified file in Bug 692216 shows a division by zero error when
filling a trapezoid with zero height. The fix is simply to update
the code not to draw zero height traps.

This should be safe as any fill adjustment is done well before this point.

The cluster shows just one difference. Looking at it in a bmpcmp it does
look suspicious, but it's not clear whether it's a progression or a
regression. Ken informs me he's seen the same differences before with
unrelated changes, so perhaps it's an indeterminism.

The code seems more sensible this way round anyway.