Bug 692218: Incorrect CIDtoGIDMap length causing crash.
In pdfwrite's pdf_convert_truetype_font_descriptor() function,
the lengths of both the CIDtoGIDMap and the CIDSet were being set
to the number of in use CID's in the font, whereas they need to
be, and were being filled, as being LastChar + 1 entries long.

As well as memory corruption, the contents would have been wrong,
too.

No cluster differences expected.