Fixes in ICC profile handling and device parameters
This fixes some issues that were encountered when setting ICC
profiles on the command line.  One issue is that the put_profile
procedure needs to be initialized a bit earlier which required
a change in gdevprn.h  In addition, when the output profile is
specified to be the CIELAB profile and there is transparency in
the file there was a rc issue with the profile when going through
the clist.  This also includes code to handle the rendering intent device
parameters for object types.