Bug #692202 Bug in display_raster function
The calculation of bytewidth, the number of bytes required for a row of
pixels, only worked correctly when the colour depth was a multiple of 8.

Since the device is a display device, and has since become used as a
general method for getting data out of GS, this is not reasonable
(monochrome and greyscale devices are less than 8 bits)

Modified as suggested by Martin Osieka to round up the calculation.