mem_mono_copy_mono optimisations.
Working on Norberts latest test files, I have changed mem_mono_copy_mono
to only use copy_rop if the runs are large enough for it to get a
benefit. Currently this threshold is set to 32 pixels (1 word).

This is sufficiently high that we get 9.01 or better performance on all the
test files, but might benefit from further tuning later.