Fix for Bug 692217 nested trans_pattern_clist SEGV
This was caused by a clip device created for tiling a mask with a transparent
pattern rendered with a clist. The clip device used for rendering did not
get its color_info updated when its target (pdf14) changed the color_info,
specifically the depth.

Fixed by grabbing the target color_info after calling the target's compositor.