Use correct enumerator for PCL type 3 fonts in pdfwrite
Part of the change for type 3 fonts from PCL used the wrong enumerator.
The code relies on a text enumerator actually being a show enumerator,
which is guaranteed true if we use the right one, but may not be if we
use the wrong one.

This didn't seem to be causing any problems, but fixed anyway.