Fix Bug 692225. Interpolate and Decode don't play nice.
Bug 692225 gives an example where an Image with both a Decode array
and interpolation set to true gives bad results.

The problem was with a branch of the if in image_render_interpolate_icc
that was assuming that any decode array given was [0 1]. The fix here is
simply to test for whether need_decode is set.

No changes seen on clusterpushing.

Check this with Michael quand il sont retourne.