Fix for issues in the setting of the newer ICC device parameters
This fixes issues that were introduced when I added the new device
parameters for rendering intent.  Also, this adds in a new structure
for defining different ICC profiles for the SOURCE colors
of objects (e.g. text image graphic).  This will be set as a user
parameter and stored in the icc manager.   Next step in this is
to get things initialized and then use the associated profiles.