Fix bug 691152; gradients appear 'stepped'.
The original bug report was that tiffsep and tiff24nc give noticably
different results. This turns out to be because tiffsep is set to use
non 'linear-and-separable' colours.

When shadings are rendered, they are 'decomposed' down into smaller
and smaller regions, until they are smaller than a given threshold.
These regions are then filled either with linear filled traps (in the
linear and separable case) or with constant colour (in the non linear
and separable case). This means that we can get away with a larger
threshold in the linear and separable case (tiff24nc) and still see
smooth results.

The original version of the code (pre SVN rev 7936) used to decompose
until the regions were smaller than a pixel. At rev 7936 this was
upped to be 1 point (1/72 of an inch) (or 1 pixel, whatever was
larger). Reverting to this original code solves the problem by making
the constant color regions small enough so that they still look
smooth.

There would therefore appear to be several possible approaches to
solve the problem:

1) We could back out the change in 7936. This trades time (27% in the
worst measured test case) for correctness.

2) We could change tiffsep to use linear and separable colours.

I've opted for an approach based on 1) here, because a bmpcmp clearly
shows easily visible differences in some files in other devices, for
example:

  tests_private/comparefiles/470-01.ps.pgmraw.300.0

By default, the code now operates as pre revision 7936. Should the
change in speed be considered unacceptable, gs can be build with the
symbol  MAX_SHADING_RESOLUTION defined to maximum dpi to which shadings
should be decomposed; building with -DMAX_SHADING_RESOLUTION=72 will
give the same results as the existing code.