Fix bug 692226; stray pixels in skewed masked image
When painting a masked image, we first plot a scanlines worth of
mask pixels to a mask plane. This is then used in a clipping device to
clip the image pixels that follow thereafter.

In the code that plots the masked pixels it currently gathers up 'runs'
of identical pixels and plots them all at once. This works fine for
portrait and landscape images, but for skewed ones has problems due to
rounding errors. By plotting large runs of pixels at once, we can a) get
gaps between subsequent rows of masked pixels, and b) get a mismatch
between the pixels covered by the mask and the image. These manifest
as holes in the image.

The image code already has a comment in it to the effect that we cannot
amalgamate large blocks due to rounding errors.

This fix therefore extends this policy (of not amalgamating) to skewed
masked images too.

426 non-pdfwrite/pswrite differences expected. 79 pdfwrite. 31 ps2write.
Checked with bmpcmp and all seem either progressions or neutral.