Bug 691222: Make windows build use UTF-8 encoding.
We change the windows builds to use the 'wmain' rather than 'main'
entrypoints. This means we get the command line supplied in 'wchar_t's
rather than chars. We convert back to chars using UTF-8 encoding, and
call (what was) the main entrypoint.

This means that we can cope with unicode filenames/paths etc.

To match the encoding, we therefore need to wrap every use of the
filenames with the associated utf-8 -> wchar_t conversion and use
the unicode file access functions (_wfopen instead of fopen etc)
instead.

Simple testing seems to indicate that this works. I think I've got
every occurence of file access, but it's possible I've missed some. If so
I'll fix them piecemeal as they are reported.

This should solve bug 691222, and hopefully 691117.