Fix Bug 689737: antialiasing issues with patterns.
Bug 689737 shows mismatches between rendering with and without
antialiasing when PaintType 2 patterns are used.

Following Michaels analysis of the bug, the fix seems to be very simple;
when opening a pattern accumulator device, if the PaintType is 2, then
set color_info.anti_alias.graphics_bits to 1. We also set
color_info.anti_alias.text_bits to 1, even though this is unlikely to
make a difference if freetype is used.

No expected cluster differences as antialiasing isn't used in any
cluster tests.

This commit also removes a FIXME (as I've checked with Michael), and
improves the commenting on the device proc structure (which I'd done
when trying a different device proc based fix, but is an improvement
anyway).