Bug 691118: Fix radial shading bug
Bug 691118 describes a bug with radial shadings in a PDF file.

Back in r7982 (git 69a43b0) (16 May 2007) Igor introduced some cunning
code to speed radial shadings. It would spot that the shading was large
in comparison with the area being filled, and produce a smaller,
equivalent shading.

Essentially, it would take the destination rectangle, and figure out
where in the shading the 'extremes' of that rectangle would come from
(allowing for distortion in shape generation) and ensure that the newly
generated shading covered the same area.

Either the corners of the rectangle could be extremes, or points on the
side. The code for the points on the sides was broken due to a typo
causing the X and Y cases to be switched.

All that is required to solve the bug is to change the == to !=.

In researching the problem however, I recommented lots of the code to
make it more comprehensible (to me at least!) along with introducing
some temporary variables to make it clearer and avoid needless
recomputation of values. I am committing this as I believe it's a step
forward for the code, even though it doesn't actually make a
difference.

48 differences in non-pdfwrite/ps2write tests. 11 in pdfwrite, 17 in
ps2write. Checked with bmpcmp, all unnoticable to the naked eye, so
presumably progressions.