Work toward getting fast thresholding for color output devices and color input images
To make my life easier, this updates msvc.mak to include the planar devices.
Changes in image_render_mono_ht to work towards the case where the output device is not mono.
Minor fixes in gxicolor.c in the portion of code that steps through the planes.
Addition of copy_plan operations in gxht_thresh.c for CMYK planar devices.
Addition of clip_copy_plane for the clip device (thanks Robin Watts).