More work on unicode handling under windows.
Continuing from commit 0ea739147, fix other issues, including:

 * OpenPrinter calls changed to use OpenPrinterW.
 * FindFirstFile/FindNextFile changed to use FindFirstFileW/FindNextFileW.
 * gp_getenv windows specific registry lookups changed to use unicode
   versions.
 * GetProfileString changed to use GetProfileStringW.
 * Add unicode changes to gswin32 as well as gswin32c.exe.

This hopefully concludes the bulk of the work on unicode support. Possible
areas to consider in future include:

 * changing gswin32 so that cut and pasting operates in unicode
 * make gswin32 display text in unicode
 * modify stdin to go from unicode to utf8, and stdout to go from utf8
   back to unicode.

I hope no idea if any of those are possible or not.

Testing seems to indicate that this all works OK, but I am not ideally
set up here.

For safety, I have introduced a WINDOWS_NO_UNICODE define. If this is
predefined during a build then this will cause the code to drop back to
the old mode of operation.